test: fix flaky hocuspocus persistence test and run orphaned chat tests under vitest
Two pre-existing test failures, unrelated to feature work:

- services/hocuspocus persistence.vitest.ts: "loads the latest snapshot
  plus the current-state row" built the snapshot and the live update row
  as two independent Y.Docs. With distinct random client ids, Yjs resolves
  the conflicting `content` writes by client id rather than causal order,
  so the assertion (live-state should win) passed or failed at random. Model
  production reality instead: the update row is a causal descendant of the
  snapshot (same client, higher clock), so it deterministically supersedes it.

- packages/chat mentionParser.test.ts / adapterUtils.test.ts were hand-rolled
  tsx console runners (custom test()/expect(), only runnable via `npx tsx`).
  Under vitest they failed with "No test suite found" since they register no
  vitest cases. Convert both to real vitest tests, rename to the repo's
  `*.vitest.ts` convention, and add a vitest config + `test` script + devDep
  to @gruenerator/chat so they actually run in CI (turbo run test).
